Separate NetLogReader into its own file
